Edit ChangeYearModal use typescript
